# [Arm64] Treat methods of non-generic Vector64 and Vector128 classes as intrinsics