@echesakovMSFT 2020-08-06
@davidwrighton Just to clarify things a bit that the changes are not related to our discussion in #32714. The methods were enabled to be treated as intrinsics on Arm64 a while ago under assumption that AdvSimd is required baseline (in #38060).